requirements.txt
async-tkinter-loop==0.9.3
customtkinter==5.2.2
darkdetect==0.8.0
evdev==1.7.0
MouseInfo==0.1.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
six==1.16.0
typing_extensions==4.10.0
